Remove `-chart` from subchart references (#2)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1800

Description of changes:
The `Chart.yaml` file uses aliases to refer to the subcharts that specifically remove the need for `-chart` suffixes in `values.yaml`. This pull request updates the README to use the direct service name when overriding `values.yaml`, and updates the default `values.yaml` with the correct subchart keys.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.